# back-api-rest/openapi.yaml
openapi: 3.0.0
info:
  title: FeYConsuelo REST Service
  description: Backend para gestión de Fe y Consuelo
  contact:
    name: Eduardo Sanchez Bonilla
    email: eduardosanchezbonilla@gmail.com
  license:
    name: Apache License Version 2.0
  version: 1.0.0
servers:
  - url: http://localhost:8082
    description: Generated server url
...
paths:
...
  "/survey":
    post:
      tags:
        - survey-controller
      summary: Add a new survey
      operationId: insertSurvey
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SurveyDto"
        required: true
      responses:
        '201':
          description: Created
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
    get:
      tags:
        - survey-controller
      summary: Returns all the surveys
      operationId: getAllSurveys
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/SurveyDto"
        '204':
          description: No content
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  "/survey/{surveyId}":
    put:
      tags:
        - survey-controller
      summary: Update any data for a Survey
      operationId: updateSurvey
      parameters:
        - name: surveyId
          in: path
          description: Survey Id
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SurveyDto"
        required: true
      responses:
        '200':
          description: OK
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
    get:
      tags:
        - survey-controller
      summary: Returns the survey
      operationId: getSurvey
      parameters:
        - name: surveyId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/SurveyDto"
        '204':
          description: No content
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
    delete:
      tags:
        - survey-controller
      summary: Delete a survey
      operationId: deleteSurvey
      parameters:
        - name: surveyId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
...
components:
  schemas:
...
    SurveyDto:
      description: Survey data
      required:
        - name
        - type
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Identifier
          example: 1
        name:
          type: string
        type:
          type: string
        description:
          type: string
        question:
          type: string
        isPublic:
          type: boolean
        isOpen:
          type: boolean
        isFinished:
          type: boolean
        image:
          type: string
        createdTime:
          type: string
        options:
          type: array
          items:
            $ref: '#/components/schemas/SurveyOptionDto'
        userOptionsVote:
          type: array
          items:
            $ref: '#/components/schemas/SurveyOptionVoteDto'
        numberUsersVote:
          type: integer
    SurveyOptionDto:
      description: Survey option data
      required:
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Identifier
          example: 1
        name:
          type: string
        description:
          type: string
        youtubeId:
          type: string
        order:
          type: integer
          example: 1
...
    SurveyOptionVoteDto:
      description: Survey option data
      required:
        - id
        - score
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Identifier
          example: 1
        score:
          type: integer

    ######################################################################
    ErrorDto:
      required:
        - message
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: '400'
        name:
          type: string
          description: Error name
          example: 'Bad Request'
        message:
          type: string
          description: Error message
          example: 'Bad Request'